drtrrtdhdfh

+++ admin_login/crud/action_insert.php
@@ -600,7 +600,7 @@ if(! isset($_SESSION['user']) && $_SESSION['user']==null){
                         function convert($str, $num)
                         {
 
-                            $str_int= (int)$str;
+                            $str_int= (float)$str;
 
                             return round($str_int,$num);
